+++ internal/repository/auth/model/model.go
@@ -8,7 +8,7 @@ type User struct {
 	Name         string    `db:"name"`
 	Email        string    `db:"email"`
 	PasswordHash string    `db:"password_hash"`
-	RoleID       int64     `db:"role_id"`
+	RoleID       int32     `db:"role_id"`
 	CreatedAt    time.Time `db:"created_at"`
 	UpdatedAt    time.Time `db:"updated_at"`
 }
